Add files via upload
This is the Admin Task Dashboard which was a task assigned by the Admin department to me. However, the main objective for creating this one is to save the time of reporting by individuals and make the work easy for the department.
The perfect and correct KPIs and stats would be delivered on time so the relevant departments got to know about the team performance and stakeholders, managers and team leads too.

So the problems the department was facing are
1 The stats and metrics were not delivered on time.
2 KPIs were delivered with some human error.
3 Department leads managers stakeholders CEO, CFO, and CFO used to wait for the stats 
4 Low team productivity and fair performance of the department

This dashboard is responsible to tell the KPIs and stats of the employees in the Admin department whether the task has been completed or not according
to the date, slicers have been selected which means it will get the complete and incomplete status of the employees about their tasks.
after this dashboard had been created now it is so helpful and supportive for the seniors to take their decisions on some serious matters to increase the team productivity. 
the dashboard is live on the Power BI service so they can monitor the team and department performance.