Add files via upload
This is great, really important, and helpful known as the Appointment Adherence dashboard that was assigned to me. so as the organization providing IT services to the UK. some engineers do their job 
and the deployment department wants the report of engineers from the MIS department to just get the number of jobs deployed in terms of these deployed jobs, the deployment 
team wants adherence timely according to that time they measure the performance of the department and engineers. when the job is created and when the engineer has arrived on that job that is the adherence.

So the data is coming from the SQL Database already connected with Power BI and based on the logic interpret by the department, I designed this dashboard. It was not seen by only the managers and group leads 
but also the managers, Head of operations, COO, CFO, and evens get monitored by the directors of the organization to see the performance of this department.

this dashboard is live on the powerBI service on the cloud so it enables the stakeholders, COOs, CFOs all senior members consistently monitor the department's performance by getting the KPIs from this dashboard and make the important decisions to increase the business operation and team productivity.

this helped the department to get rid of designing the report on Excel by 6 times a day and saved time too, make easier for the senior members to take the decision more quickly
based on this dashboard KPIs metrics and stats.
The Chief head of the operation applaud and admired this dashboard which was a very great moment for me. The managers also use this dashboard to get represented in executive meetings.

This dashboard is answering these questions as storytelling.
- How many jobs are deployed and done in each specific region?
- What are the statuses of jobs being deployed to engineers in each region?
- How many engineers arrived in 24 hours, under 24-48 hours and over 48 hours?