Add files via upload
Claims Dashboard is the most valuable dashboard for the senior members (Team leads, Managers, CFOs COO). Before this dashboard, the FNOL claims department sent four to six reports weekly and daily.

This dashboard defines the claims getting from the customer by the organization. so it interprets all about the claims like claims accepted,
claim received (Boiler, Home emergency, Appliances, SW) softix (fixed the problem too quickly) and claim declined.

After this dashboard now it's too quick for all the seniors to take decisions and it also helped the department to get to know about the performance.

A board pack meeting is a combo pack of 13 to 15 reports that takes 8 to 9 hours to make and the reports are then represented in the meeting to summarize across the departments in which the senior member used to attend.

This dashboard is one of the most important dashboards represented in the Board meeting. now this dashboard is live on Power BI service running on, monitored and connected to real-time data SQL Server Database.