Add files via upload
Daily Deployment Dashboard is the dashboard created by me requested by the Deployment department to see how many jobs are assigned to employed, subcontractor, PAYgs and Manufacturers engineers daily monthly weekly and yearly basis.

This is one of the dashboards represented in a board pack meeting on weekly basis to see where the company stands in assigning and completing the jobs by engineers. it splits the jobs into breakdown repairs and services. 

This dashboard is used by the leads, managers, and senior and upper-level management like stakeholders, COO, CFO and CEO to know the performance of the department. it's been automated and running on real-time data with just a single click of schedule refresh. I also implemented and integrated this dashboard with power automate to get job info data with just a single click by an AI bot named power automate which usually sends on Microsoft Teams to whoever clicks that button.