Add files via upload
First visit fix dashboard that defines the engineer's performance according to their work-related it includes the employed and subcontractors as well. It tell the recalls and FVF status that how many engineer fix the problem of the customer on first visit known as FVF.
Recalls means the customers calls and company assigned the job to engineer to fix the issue of customer engineer fixed the issue but next time the same issue was raised with in 2 to 3 days counted as fault of engineer. this summarizes the engineer's performance.

On the basis of this dashboard the head of operation takes the decision according to the results on weekly and monthly basis.
before this dashboard 4 to 7 reports were sent across the departments.
but now after dashboard automation head of operation can reach directly to the metrics and result to judge the outcome wihtout any request.

This one is also represented in board meetings.