Add files via upload
Another dashboard named Engineer Scorecard was requested by the COO and CFO of the company and this one is a mega and unique project for me to work on. this took me 3 months to create and design. It's the combination of 3 databases and some live sheets.

It generally summarizes all the information about the services and operations that 4 to 5 departments provide and serves the customers. this dashboard owns by the COO and CFO to consistently know about the outcomes and performance of the Engineers consists of these aspects.
Complaints
services
job deployed
FTE (Attendance of engineer to monitor how many jobs he has done)
sales and leads
Now after the automation of this dashboard, the head of the operation doesn't need separate individual bulk of reports to take actions on engineer performance in fact its live on Power BI service to just only select an engineer and get the selected engineer report afterwards schedule one to one session in which head of the operation discuss matters in meeting with Engineer to talk about the performance.
the data can also be extracted about the engineers according to the filters selected.