Add files via upload
This one is FNOL Dashboard that is linked to the claims department it tells the department about the claims they received with their statuses. this one was been monitored by the manager of the claims department. and it's also automated so the relevant manager gets the updated data from the SQL server in every two Hours.