Add files via upload
Another dashboard that was designed is Issuance Dashboard which is related to another department. this report was on Excel before the automated dashboard.
This dashboard measures the Engineer's performance on the basis of their attendance and jobs deployed on the behalf of which the percentage was calculated and sent to seniors 6 to 7 times a day.

So to prevent this I transferred to the power Bi that is running on real-time data that reduces the time and human error as well. So this is being monitored by senior members of the company.